refactor(engine): extract generation options + KV budget from inference.py
Move two leaf clusters out of inference.py (5,497 → ~5,050 lines):

- generation_options.py: _merge_default_options, _apply_sampling_defaults
  (Ollama-parity defaults, #646), _build_generate_kwargs, _apply_seed
- kv_budget.py: MEMORY_SAFETY_FACTOR, estimate_kv_cache_bytes,
  tokenize_for_cache, build_context_input_tokens, _parse_kv_cache_quant

inference.py re-exports every moved name so imports and facade patch
targets for facade-resident consumers are unchanged; tests that patch a
moved consumer's dependencies (_apply_seed's mx, _build_generate_kwargs'
make_sampler/make_logits_processors) are repointed to generation_options.
The heavily-patched orchestration (streaming/batched/cache-setup paths,
aux generate_*) deliberately stays put — moving it invalidates ~200
facade patch sites and needs its own test-migration pass (follow-up).
